pyproject.toml
[project]
name = "Substrate"
version = "0.1.0"
description = "test"
authors = [{ name = "Mike Waters", email = "mike@mikewaters.net" }]
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "marimo==0.15.5",
    "pydantic-settings>=2.11.0",
    "loguru>=0.7.0",
    "mw",
    "bitsandbytes@{root:uri}/src/bitsandbytes",
    "advanced_alchemy@{root:uri}/src/advanced-alchemy",
]

[project.scripts]
#validate-schema = "ontology.domains:validate"
ontology = "ontology.cli.main:main"

[tool.uv]
package = true

[tool.marimo.runtime]
pythonpath = ["src"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "src/ontology",
    "src/index",
    "src/mw",
    "src/advanced_alchemy",
    "src/bitsandbytes",
    ]

[dependency-groups]
ontology = [
    "polars>=1.33.1",
    "pydantic>=2.11.9",
    "owlready2>=0.48",
    "dotenv>=0.9.9",
    "linkml>=1.9.4",
    "logzero>=1.7.0",
    "cymple>=0.12.0",
    "terminusdb-client>=10.2.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "fastapi[standard]>=0.119.1",
    "httpx>=0.28.1",
    #"typer>=0.20.0",
    "rich>=14.2.0",
    "pydantic-settings>=2.11.0",
    "aiosqlite>=0.21.0",
    "pydantic-ai>=1.9.1",
    "uvicorn[standard]>=0.30.0",
    "txtai[graph,pipeline]>=9.1.0",

]
index = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "llama-index>=0.12.0",
    "loguru>=0.7.0",
    "sqlalchemy>=2.0",
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-obsidian",
    "llama-index-embeddings-huggingface>=0.6.1",
    "mlx-embeddings>=0.0.5",
    "llama-index-vector-stores-duckdb>=0.5.1",
    "llama-index-storage-docstore-duckdb>=0.2.1",
]
dev = [
    "basedpyright>=1.33.0",
    #"basedpyright>=1.31.7",
    #"browser-use[cli]>=0.8.0",
    #"hypothesis>=6.138.16",
    "ipdb>=0.13.13",
    "mdtree-mermaid==0.8",
    "pudb>=2025.1.1",
    "pyfakefs>=5.9.3",
    "pylsp-rope>=0.1.17",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "pytest-testmon>=2.1.4",
    "python-lsp-server[all]>=1.13.1",
    "pyyaml>=6.0.3",
    "mypy",
    "ruff",
    "tach",
    "respx",
]

###
### Testing
###
[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "-v",
    "--log-cli-level=INFO",
    #"--testmon",
    #"--cov=src/index --cov=src/ontology --cov-report=term-missing --cov-report=html",
    "--ignore=src/index/tests/scenario"
]
testpaths = [
    "src/ontology",
    "src/index",
    #"tools"  #FIXME
]
norecursedirs = "src/advanced-alchemy src/aiosqlite"
pythonpath = [
  "."
]
env = [
    "PYTHONBREAKPOINT=ipdb.set_trace",
    "SUBSTRATE_DATABASE_ECHO=true",
    "SUBSTRATE_DATABASE_DB_PATH=.data/yaaaaaa.db",
    "SUBSTRATE_ENVIRONMENT=test"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow",
    "requires_llm: tests requiring LLM provider",
]

[tool.pytest.env]
PYTHONBREAKPOINT = "ipdb.set_trace"
SUBSTRATE_DATABASE_ECHO = 'true'
SUBSTRATE_DATABASE_DB_PATH=".data/yaaaaaa.db"
SUBSTRATE_ENVIRONMENT = "test"

[tool.coverage.run]
source = ["src/index", "src/ontology"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true

###
### Linting
###

[tool.ruff]
# Ruff already excludes anything in .gitignore (pyright does not)
exclude = [
    # generated by LinkML
    'src/ontology/domains/database_models.py',
    'src/ontology/domains/schema_dataclasses.py',
    'src/ontology/domains/schema_models.py',
    # generated by Marimo
    'notebook.py',
    # copied from elsewhere, dont give a shit
    'src/ontology/domains/export_cypher.py',
    'patches/',
    # local stash of old code
    'meatspace/',
    # this project is not responsible for maintaining my claude framework
    '.claude/', 'tools/', 'context/',
    # other shite
    'src/aiosqlite', 'src/advanced-alchemy'
]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
ignore = [
    # pdb breakpoints
    'E702', 'I001',
    # pedantic
    'E501',  # line too long
    # temp
    'F841',
    'E402',
    'F401'
]
# select = [
#     # pycodestyle
#     "E",
#     # Pyflakes
#     "F",
#     # pyupgrade
#     "UP",
#     # flake8-bugbear
#     "B",
#     #"SIM", #TODO
#     # isort
#     "I",
# ]
[tool.pyright]
exclude = [
    '.venv',
    # staaahp being so annoying
    'src/ontology/tests/', 'src/ontology/domains/tests/',
    'scripts/rdf-entity-lookup.py',

    ## from Ruff:
    # generated by LinkML
    'src/ontology/domains/database_models.py',
    'src/ontology/domains/schema_dataclasses.py',
    'src/ontology/domains/schema_models.py',
    # generated by Marimo
    'notebook.py',
    # copied from elsewhere, dont give a shit
    'src/ontology/domains/export_cypher.py',
    'patches/',
    # local stash of old code
    'meatspace/',
    # this project is not responsible for maintaining my claude framework
    '.claude/', 'tools/', '.data/', 'ai_context/', 'ai_working/',
    # other shite
    'src/aiosqlite', 'src/advanced-alchemy'
]
# This is a `basedpyright`-specific config, but i get bitched at if I have
# both sections in the toml; hilariously, Basedpyright is the one that complains
# [tool.basedpyright]
#typeCheckingMode = 'standard'
typeCheckingMode = 'off'

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true

